Move UnsupportedMessageException to protocol.common package
