--- src/main/java/com/streamr/client/exceptions/UnsupportedMessageException.java
@@ -1,9 +1,0 @@
-package com.streamr.client.exceptions;
-
-public class UnsupportedMessageException extends RuntimeException {
-
-    public UnsupportedMessageException(String message) {
-        super(message);
-    }
-
-}
