+++ src/main/java/com/streamr/client/protocol/common/UnsupportedMessageException.java
@@ -1,0 +1,7 @@
+package com.streamr.client.protocol.common;
+
+public class UnsupportedMessageException extends RuntimeException {
+  public UnsupportedMessageException(final String message) {
+    super(message);
+  }
+}
