+++ src/main/java/com/streamr/client/subs/BasicSubscription.java
@@ -3,7 +3,7 @@ package com.streamr.client.subs;
 import com.streamr.client.MessageHandler;
 import com.streamr.client.exceptions.GapDetectedException;
 import com.streamr.client.exceptions.UnableToDecryptException;
-import com.streamr.client.exceptions.UnsupportedMessageException;
+import com.streamr.client.protocol.common.UnsupportedMessageException;
 import com.streamr.client.protocol.common.MessageRef;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.utils.Address;
