+++ src/main/java/com/streamr/client/subs/CombinedSubscription.java
@@ -3,6 +3,7 @@ package com.streamr.client.subs;
 import com.streamr.client.MessageHandler;
 import com.streamr.client.exceptions.*;
 import com.streamr.client.options.ResendOption;
+import com.streamr.client.protocol.common.UnsupportedMessageException;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.utils.*;
