+++ src/main/java/com/streamr/client/subs/RealTimeSubscription.java
@@ -2,17 +2,14 @@ package com.streamr.client.subs;
 
 import com.streamr.client.MessageHandler;
 import com.streamr.client.exceptions.GapDetectedException;
-import com.streamr.client.exceptions.UnableToDecryptException;
-import com.streamr.client.exceptions.UnsupportedMessageException;
+import com.streamr.client.protocol.common.UnsupportedMessageException;
 import com.streamr.client.options.ResendOption;
 import com.streamr.client.protocol.message_layer.StreamMessage;
 import com.streamr.client.utils.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.crypto.SecretKey;
 import java.util.ArrayList;
-import java.util.HashSet;
 
 public class RealTimeSubscription extends BasicSubscription {
